move CommandStackManager out of internal
