+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/di/DiagramModule.java
@@ -87,9 +87,9 @@ import org.eclipse.glsp.server.gmodel.GModelCutOperationHandler;
 import org.eclipse.glsp.server.gson.GraphGsonConfigurationFactory;
 import org.eclipse.glsp.server.internal.actions.DefaultActionDispatcher;
 import org.eclipse.glsp.server.internal.actions.DefaultActionHandlerRegistry;
-import org.eclipse.glsp.server.internal.command.CommandStackFactory;
-import org.eclipse.glsp.server.internal.command.CommandStackManager;
-import org.eclipse.glsp.server.internal.command.DefaultCommandStackManager;
+import org.eclipse.glsp.server.command.CommandStackFactory;
+import org.eclipse.glsp.server.command.CommandStackManager;
+import org.eclipse.glsp.server.command.DefaultCommandStackManager;
 import org.eclipse.glsp.server.internal.diagram.DefaultServerConfigurationContribution;
 import org.eclipse.glsp.server.internal.featues.directediting.DefaultContextEditValidatorRegistry;
 import org.eclipse.glsp.server.internal.featues.navigation.DefaultNavigationTargetProviderRegistry;
