+++ plugins/org.eclipse.glsp.server/src/org/eclipse/glsp/server/internal/gmodel/commandstack/GModelCommandStackFactory.java
@@ -1,7 +1,7 @@
 package org.eclipse.glsp.server.internal.gmodel.commandstack;
 
 import org.eclipse.emf.common.command.CommandStack;
-import org.eclipse.glsp.server.internal.command.CommandStackFactory;
+import org.eclipse.glsp.server.command.CommandStackFactory;
 
 public class GModelCommandStackFactory implements CommandStackFactory  {
     @Override
